# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/housing-units/{unitId}/rooms": {
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a room in a unit (owner only)",
        "tags": [
          "rooms"
        ]
      },
      "get": {
        "operationId": "RoomsController_findAll",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoomDto"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "List rooms in a unit",
        "tags": [
          "rooms"
        ]
      }
    },
    "/housing-units/{unitId}/rooms/{roomId}": {
      "get": {
        "operationId": "RoomsController_findOne",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Get a room",
        "tags": [
          "rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RoomDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update a room (owner only)",
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "RoomsController_remove",
        "parameters": [
          {
            "name": "unitId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roomId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Delete a room (owner only)",
        "tags": [
          "rooms"
        ]
      }
    },
...
  },
  "info": {
    "title": "RoomMate API",
    "description": "Shared-housing room reservation API",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "maxReservationHours": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "maxReservationsPerDay": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "minGapMinutes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "nullable": true
          },
          "availableFrom": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "nullable": true
          },
          "availableTo": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "nullable": true
          }
        },
        "required": [
          "name"
        ]
      },
      "RoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "unitId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "maxReservationHours": {
            "type": "integer",
            "nullable": true
          },
          "maxReservationsPerDay": {
            "type": "integer",
            "nullable": true
          },
          "minGapMinutes": {
            "type": "integer",
            "nullable": true
          },
          "availableFrom": {
            "type": "string",
            "nullable": true
          },
          "availableTo": {
            "type": "string",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "unitId",
          "name",
          "description",
          "maxReservationHours",
          "maxReservationsPerDay",
          "minGapMinutes",
          "availableFrom",
          "availableTo",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "nullable": true
          },
          "maxReservationHours": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "maxReservationsPerDay": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": true,
            "nullable": true
          },
          "minGapMinutes": {
            "type": "integer",
            "minimum": 0,
            "exclusiveMinimum": false,
            "nullable": true
          },
          "availableFrom": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "nullable": true
          },
          "availableTo": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "nullable": true
          }
        }
      },
...
    }
  }
}